+++ components/assets/styles/StylesKKB.js
@@ -1,0 +1,23 @@
+import { StyleSheet } from "react-native";
+const styles = StyleSheet.create({
+    boxes:{
+      width:300,
+      backgroundColor:'#000080',
+      height:70,
+      borderRadius:10,
+      justifyContent:'center',
+      alignItems:'center'
+    },
+    text_out:{
+        fontSize:28,
+        fontWeight:'bold',
+        color:'black'
+    },
+    text_in:{
+      fontSize:25,
+      fontWeight:'bold',
+      color:'white'
+    }
+  })
+
+export default styles
