+++ components/firebase/firestoreContants.js
@@ -1,0 +1,3 @@
+export const FIRESTORE_COLLECTION_PATHS = {
+    'Saadhana' : 'saadhana' 
+}
