+++ package.json
@@ -25,6 +25,7 @@
     "fbjs": "^3.0.5",
     "react": "18.2.0",
     "react-native": "0.74.5",
+    "react-native-chart-kit": "^6.12.0",
     "react-native-gesture-handler": "^2.19.0",
     "react-native-image-picker": "^7.1.2",
     "react-native-modal-datetime-picker": "^18.0.0",
@@ -32,6 +33,7 @@
     "react-native-safe-area-context": "^4.11.0",
     "react-native-screens": "^3.34.0",
     "react-native-snap-carousel": "^3.9.1",
+    "react-native-svg": "^15.7.1",
     "react-native-swiper": "^1.6.0",
     "react-native-webview": "^13.12.2"
   },
